+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/preference/TogglePreference.java
@@ -5,6 +5,7 @@
 
 package app.morphe.extension.tiktok.settings.preference;
 
+import app.morphe.extension.tiktok.settings.L10n;
 import android.content.Context;
 import android.preference.SwitchPreference;
 import android.view.View;
@@ -29,5 +30,14 @@ public class TogglePreference extends SwitchPreference {
 
         Utils.setTitleAndSummaryColor(view);
     }
-}
 
+    @Override
+    public void setTitle(CharSequence title) {
+        super.setTitle(L10n.t(getContext(), title));
+    }
+
+    @Override
+    public void setSummary(CharSequence summary) {
+        super.setSummary(L10n.t(getContext(), summary));
+    }
+}